generic crawl controller
Moved CrawlControler logic into a parametrizable version, GenericCrawlControler, which allows to  specify the minimum WebCrawler class allowed.

Declared protected Object collectCrawlerData(CrawlerType crawler) and called it inside start to extract data from crawler, so subclases can override it for custom data extraction based on WebCrawler type.
